// src/models/requests/Users.Requests.ts
import { JwtPayload } from 'jsonwebtoken'
import { TokenType } from '../../constants/enums'
import { ParamsDictionary } from 'express-serve-static-core'
export interface RegisterReqBody {
  name: string
  email: string
  password: string
  confirm_password: string
  date_of_birth: string
}
export interface LogoutReqBody {
  refresh_token: string
}
export interface TokenPayload extends JwtPayload {
  user_id: string
  token_type: TokenType
}
export interface forgotPasswordReqBody {
  email: string
}
export interface LoginReqBody {
  email: string
  password: string
}
export interface ResetPasswordReqBody {
  password: string
  confirm_password: string
  forgot_password_token: string
}
export interface VerifyForgotPasswordReqBody {
  forgot_password_token: string
}
export interface UpdateMeReqBody {
  name?: string
  date_of_birth?: string
  bio?: string
  location?: string
  website?: string
  username?: string
  avatar?: string
  cover_photo?: string
}
export interface followReqBody {
  followed_user_id: string
}
export interface GetProfileReqPanams extends ParamsDictionary {
  username: string
}
export interface UnfollowReqParams extends ParamsDictionary {
  user_id: string
}
export interface ChangePasswondReqBody {
  old_password: string
  password: string
  confirm_password: string
}